Renamed django app
