+++ backend/game_service/game_logic/routing.py
@@ -2,5 +2,5 @@ from django.urls import path
 from .consumers import GameConsumer
 
 websocket_urlpatterns = [
-    path('ws/game/<int:lobby_id>/', GameConsumer.as_asgi()),
+    path('ws/game/<int:game_id>/', GameConsumer.as_asgi()),
 ]
