--- backend/game_service/game_socket/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class WebsocketsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'game_socket'
