+++ frontend/game/game_logic.js
@@ -193,13 +193,11 @@ const startMatchmaking = () => {
         console.log("Message received: ", data);
         if (data.type === 'match_found') {
             const gameId = data.message.game_id;
-
-            if (gameId) {
-                console.log("Match found with game ID:", gameId);
-                // startGame(gameId);
-            } else {
-                console.error("Match found but no game ID provided. Data:", data.message);
-            }
+            console.log("Match found with game ID:", gameId);
+            // proceed to start the game
+            // startGame(gameId)
+        } else if (data.status === 'waiting_in_queue') {
+            console.log("Waiting in queue for an opponent");
         }
     };
 
@@ -209,123 +207,3 @@ const startMatchmaking = () => {
 }
 
 document.getElementById('startMatchmakingBtn').addEventListener('click', startMatchmaking);
-
-// ------------------- OLD CODE:
-
-// const startMatchmaking = () => {
-//     console.log("Trying to setup connection....");
-//     const matchmakingSocket = new WebSocket('ws://127.0.0.1:8002/ws/matchmaking/');
-//
-//     matchmakingSocket.onopen = () => {
-//         console.log("Matchmaking socket connection established");
-//
-//         matchmakingSocket.send(JSON.stringify({
-//             action: 'join_queue',
-//             username: player,
-//         }));
-//     };
-//
-//     matchmakingSocket.onmessage = (e) => {
-//         const data = JSON.parse(e.data);
-//         //console.log('Message received: ', data);
-//
-//         if (data.type === 'match_found') {
-//              // console.log("Match found: ", data.message);
-//             startGame(data.message.message.id);
-//         }
-//     };
-//
-//     matchmakingSocket.onclose = function(e) {
-//         console.log('Matchmaking WebSocket connection closed');
-//     };
-// };
-//
-// const startGame = (lobbyId) => {
-//     const canvas = document.getElementById('pongCanvas');
-//     const ctx = canvas.getContext('2d');
-//     canvas.style.display = 'block';
-//
-//     const ws = new WebSocket(`ws://127.0.0.1:8002/ws/game/${lobbyId}/`);
-//
-//     ws.onopen = (event) => {
-//         console.log("Websocket is now open!", event);
-//     };
-//
-//     ws.onclose = (event) => {
-//         console.log("Websockets connection closed!");
-//     };
-//
-//     ws.onerror = (event) => {
-//         console.error("Websocket error observed: ", event);
-//     };
-//
-//     ws.onmessage = (event) => {
-//         const data = JSON.parse(event.data);
-//
-//         if (data.type === 'game_state') {
-//             updateGame(data.message);
-//             updateScores(data.message.scores);
-//         }
-//         else if (data.type == 'game_over') {
-//             console.log(data.message.winner + " has won!");
-//         }
-//     };
-//
-//     const updateScores = (scores) => {
-//         document.getElementById('player1Score').innerHTML = scores.player1;
-//         document.getElementById('player2Score').innerHTML = scores.player2;
-//     }
-//
-//     const updateGame = (gameState) => {
-//         if (ctx) {
-//             ctx.clearRect(0, 0, canvas.width, canvas.height);
-//
-//             // Draw background
-//             ctx.fillStyle = "#1b1b1b";
-//             ctx.fillRect(0, 0, canvas.width, canvas.height);
-//
-//             // Draw paddles
-//             ctx.fillStyle = "red";
-//             //console.log(gameState.paddle_data);
-//
-//             // Player 1
-//             ctx.fillRect(gameState.players_data.player1.paddle.x, gameState.players_data.player1.paddle.y,
-//                 gameState.players_data.player1.paddle.width, gameState.players_data.player1.paddle.height);
-//
-//             // Player 2
-//             ctx.fillRect(gameState.players_data.player2.paddle.x, gameState.players_data.player2.paddle.y,
-//                 gameState.players_data.player2.paddle.width, gameState.players_data.player2.paddle.height);
-//
-//             // Draw ball
-//             ctx.beginPath();
-//             ctx.arc(gameState.ball_data.position.x, gameState.ball_data.position.y, 10, 0, Math.PI * 2);
-//             ctx.fillStyle = "red";
-//             ctx.fill();
-//             ctx.closePath();
-//         } else {
-//             console.error("Canvas not found or not initialized");
-//         }
-//     };
-//
-//     document.addEventListener("keydown", (event) => {
-//         let action;
-//         if (event.code === "KeyW") {
-//             action = {
-//                 action: 'move_paddle',
-//                 player: player,
-//                 direction: 'up'
-//             };
-//         } else if (event.code === "KeyS") {
-//             action = {
-//                 action: 'move_paddle',
-//                 player: player,
-//                 direction: 'down'
-//             };
-//         }
-//         if (action) {
-//             ws.send(JSON.stringify(action));
-//         }
-//     });
-// };
-//
-// document.getElementById('startMatchmakingBtn').addEventListener('click', startMatchmaking);
